New cvar sv_entityfile
Default behaviour is to load mapname.ent file.  To allow multiple .ent
files to be supported for one map, this field can be set to the name of
the file before changing the level.

The entity file will only be loaded if it starts with the name of the map
being loaded, and will be cleared if it doesn't match the map being loaded.

This makes sure we don't cause the server to keep spawning one map's
entities after a map change.  Any mod setting sv_entityfile should be
aware of this and responsible for keeping the combination of sv_entityfile
and map correct, but this gives a bit of protection for anyone directly
typing "map x" at the console.