fix typos in cross file for the ar/c++/strip binaries
It specified the gcc frontend instead, which usually went unnoticed as
long as no static libraries were ever built.

After adding a pcre subproject wrap, the pcre library itself got built
statically, and this typo caused a build error when the ar command did
not work.

While we are here, make sure all binaries are correct, even the ones
that are still unused.