Randomize the location in svs.clients array for new clients
Players are assigned the first available spot in the svs.clients array when
they connect. On an empty server, this means that player1 is placed at
svs.clients[0], player2 at svs.clients[1], and so on.

KTX iterates over svs.clients at the start of a match to decide where each
player will spawn. This becomes problematic when there are fewer spawn points
than players. A good example is dm3, which only has six spawns.

In the scenario where players connect to an empty server, they occupy the
indexes 0 through 7 in the svs.clients array and are processed in this
predictable order. This means player1 gets a random spawn before player2,
player2 before player3, and so forth.

When it reaches player7, all spawn points are occupied. A random spawn is
selected, and the player at that spawn point will be telefragged. Similarly,
player8, the last to be processed, will also cause a telefrag.

By randomizing where new players are placed in the svs.clients array, we
eliminate the possibility of predicting where in the array a player will
be located, and thus also remove the possibility of manipulating the
telefrag outcomes.