Add message broadcasting
By setting sv_broadcast_enabled to 1, connected players can broadcast
messages to other servers by typing:

.qw <message>

The server will emit the message to all servers retrieved from the
configured master servers.

The payload that is sent is formatted as follows:
\hostport\<hostport>\name\<name>\message\<message>

When a server receives a message, the IP address and port of the sending
server are prepended to the payload in the following format:
\addr\<ip>:<port>

If \hostport is set, that information is used in the message,
otherwise, the sender's IP address and port are used as a fallback.

Example:
[tot.qwsv.net:27500] ToT_Oddjob: prac please

Messages are received using the new connectionless command broadcast,
which serves as the endpoint for accepting messages from other servers.

By default, only servers discovered through master server queries are
allowed to send messages. This validation can be disabled by setting
sv_broadcast_sender_validation_enabled to 0.

The server automatically queries the configured master servers every 10
minutes. You can manually trigger the query by executing the
updatebroadcasts command in the console.

New command:

- updatebroadcasts
  Queries the master servers and updates the list of broadcast servers.

New cvars:

- sv_broadcast_enabled
  Enables or disables broadcasting to all connected clients.

- sv_broadcast_sender_validation_enabled
  Validates the sender's address before forwarding messages to clients.