Send broadcast messages on a new socket
Previously, broadcast messages were sent using the main server socket.
Since this socket is non-blocking, it could become exhausted.

With this change, we now create a new socket for each broadcast. The
server's port number is included in the broadcast payload instead. This
allows us to remove the previously added sleep hack, which was used to
avoid flooding the main server socket.

In the previous code, when sv_broadcast_sender_validation_enabled was
enabled, we validated both the source address and port. This is no
longer possible, as the source port will now be randomly selected.
Instead, we will validate only the source address, which should be
sufficient.