Three minor bot related fixes (#177)
* Fix userinfo star usage in PF2_SetBotUserInfo

The function incorrectly assumed that the provided userinfo key wasn't a
star key.

* Display bot skill as ping

If the client is a bot and has the *skill userinfo key set, display that
value as the ping. Otherwise, fall back to the default bot ping of 10.

* Add STATUS_SHOWCLIENTTYPE to SVC_Status

Add a new bitfield, STATUS_SHOWCLIENTTYPE, that outputs an extra column
in the SVC_Status output: 'h' for humans and 'b' for bots. This could
potentially be extended in the future to support other client types,
such as service clients.