Update broadcast message (#176)
* Allow Info_ValueForKey to parse more values

Previously, only up to 4 key-value pairs could be parsed before the
oldest entries were dropped. This change increases the limit to 8
values.

* Include current player count in broadcast messages

This change adds the current number of players and the max clients value
to the broadcast.

In the SVC broadcast function, we check whether this new data is
available. If it is, a new format is used when relaying data to clients.

Old format:
[tot.qwsv.net:27500] ToT_Oddjob: prac now

New format:
> tot.qwsv.net:27500 [6/8] ToT_Oddjob: prac now